requirements.txt
pytube3
openai
youtube-transcript-api
ffmpeg-python